backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
aiosqlite==0.22.1
alembic==1.18.3
    # via backend
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via pwdlib
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.31.0
    # via backend
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==2.0.0
    # via argon2-cffi-bindings
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.128.2
    # via backend
fastapi-cli==0.0.20
    # via fastapi
fastapi-cloud-cli==0.11.0
    # via fastapi-cli
fastar==0.8.0
    # via fastapi-cloud-cli
greenlet==3.3.1 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
grpcio==1.78.0
    # via
    #   backend
    #   grpcio-reflection
    #   grpcio-tools
    #   py-grpc-prometheus
grpcio-reflection==1.78.0
    # via backend
grpcio-tools==1.78.0
    # via backend
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hiredis==3.3.0
    # via backend
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via fastapi
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via pytest
prometheus-client==0.24.1
    # via
    #   backend
    #   prometheus-fastapi-instrumentator
    #   py-grpc-prometheus
prometheus-fastapi-instrumentator==7.1.0
    # via backend
protobuf==6.33.5
    # via
    #   backend
    #   grpcio-reflection
    #   grpcio-tools
pwdlib==0.3.0
    # via backend
py-grpc-prometheus==0.8.0
    # via backend
pycparser==3.0 ; implementation_name != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   backend
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.11.0
    # via fastapi
pydantic-settings==2.12.0
    # via
    #   backend
    #   fastapi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.11.0
    # via backend
pytest==9.0.2
    # via pytest-asyncio
pytest-asyncio==1.3.0
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via fastapi
pyyaml==6.0.3
    # via uvicorn
rich==14.3.2
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.18.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk==2.52.0
    # via fastapi-cloud-cli
setuptools==82.0.0
    # via
    #   grpcio-tools
    #   py-grpc-prometheus
shellingham==1.5.4
    # via typer
sqlalchemy==2.0.46
    # via
    #   alembic
    #   backend
starlette==0.50.0
    # via
    #   fastapi
    #   prometheus-fastapi-instrumentator
typer==0.21.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   pytest-asyncio
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via sentry-sdk
uvicorn==0.40.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
